fix(antnlp->antu): fix library name.
