+++ antu/io/fields/sequence_label_field.py
@@ -1,7 +1,7 @@
 from typing import List, Iterator, Dict
 from overrides import overrides
-from pyAnt.io.token_indexers import TokenIndexer
-from pyAnt.io.vocabulary import Vocabulary
+from antu.io.token_indexers import TokenIndexer
+from antu.io.vocabulary import Vocabulary
 
 class SequenceLabelField(Field):
