+++ antu/nn/dynet/attention_mechanism.py
