+++ doc/api/antnlp.io.dataset_readers.rst
@@ -1,0 +1,22 @@
+antnlp.io.dataset\_readers package
+==================================
+
+Submodules
+----------
+
+antnlp.io.dataset\_readers.dataset\_reader module
+-------------------------------------------------
+
+.. automodule:: antnlp.io.dataset_readers.dataset_reader
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: antnlp.io.dataset_readers
+    :members:
+    :undoc-members:
+    :show-inheritance:
