+++ doc/api/antnlp.io.datasets.rst
@@ -1,0 +1,22 @@
+antnlp.io.datasets package
+==========================
+
+Submodules
+----------
+
+antnlp.io.datasets.dataset module
+---------------------------------
+
+.. automodule:: antnlp.io.datasets.dataset
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: antnlp.io.datasets
+    :members:
+    :undoc-members:
+    :show-inheritance:
