+++ doc/api/antnlp.io.fields.rst
@@ -1,0 +1,46 @@
+antnlp.io.fields package
+========================
+
+Submodules
+----------
+
+antnlp.io.fields.field module
+-----------------------------
+
+.. automodule:: antnlp.io.fields.field
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+antnlp.io.fields.index\_field module
+------------------------------------
+
+.. automodule:: antnlp.io.fields.index_field
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+antnlp.io.fields.sequence\_label\_field module
+----------------------------------------------
+
+.. automodule:: antnlp.io.fields.sequence_label_field
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+antnlp.io.fields.text\_field module
+-----------------------------------
+
+.. automodule:: antnlp.io.fields.text_field
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: antnlp.io.fields
+    :members:
+    :undoc-members:
+    :show-inheritance:
