+++ doc/api/antnlp.io.rst
@@ -1,0 +1,40 @@
+antnlp.io package
+=================
+
+Subpackages
+-----------
+
+.. toctree::
+
+    antnlp.io.dataset_readers
+    antnlp.io.datasets
+    antnlp.io.fields
+    antnlp.io.token_indexers
+
+Submodules
+----------
+
+antnlp.io.instance module
+-------------------------
+
+.. automodule:: antnlp.io.instance
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+antnlp.io.vocabulary module
+---------------------------
+
+.. automodule:: antnlp.io.vocabulary
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: antnlp.io
+    :members:
+    :undoc-members:
+    :show-inheritance:
