+++ doc/api/antnlp.io.token_indexers.rst
@@ -1,0 +1,38 @@
+antnlp.io.token\_indexers package
+=================================
+
+Submodules
+----------
+
+antnlp.io.token\_indexers.char\_token\_indexer module
+-----------------------------------------------------
+
+.. automodule:: antnlp.io.token_indexers.char_token_indexer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+antnlp.io.token\_indexers.single\_id\_token\_indexer module
+-----------------------------------------------------------
+
+.. automodule:: antnlp.io.token_indexers.single_id_token_indexer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+antnlp.io.token\_indexers.token\_indexer module
+-----------------------------------------------
+
+.. automodule:: antnlp.io.token_indexers.token_indexer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: antnlp.io.token_indexers
+    :members:
+    :undoc-members:
+    :show-inheritance:
