+++ doc/api/antnlp.nn.dynet.rst
@@ -1,0 +1,62 @@
+antnlp.nn.dynet package
+=======================
+
+Submodules
+----------
+
+antnlp.nn.dynet.attention\_mechanism module
+-------------------------------------------
+
+.. automodule:: antnlp.nn.dynet.attention_mechanism
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+antnlp.nn.dynet.char2word\_embedder module
+------------------------------------------
+
+.. automodule:: antnlp.nn.dynet.char2word_embedder
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+antnlp.nn.dynet.initializer module
+----------------------------------
+
+.. automodule:: antnlp.nn.dynet.initializer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+antnlp.nn.dynet.multi\_layer\_perception module
+-----------------------------------------------
+
+.. automodule:: antnlp.nn.dynet.multi_layer_perception
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+antnlp.nn.dynet.nn\_classifier module
+-------------------------------------
+
+.. automodule:: antnlp.nn.dynet.nn_classifier
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+antnlp.nn.dynet.rnn\_builder module
+-----------------------------------
+
+.. automodule:: antnlp.nn.dynet.rnn_builder
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: antnlp.nn.dynet
+    :members:
+    :undoc-members:
+    :show-inheritance:
