+++ doc/api/modules.rst
@@ -1,0 +1,8 @@
+antnlp
+======
+
+.. toctree::
+   :maxdepth: 4
+
+   antnlp
+   setup
