+++ doc/api/setup.rst
