--- doc/source/data.fields.rst
@@ -1,22 +1,0 @@
-data.fields package
-===================
-
-Submodules
-----------
-
-data.fields.field module
-------------------------
-
-.. automodule:: data.fields.field
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: data.fields
-    :members:
-    :undoc-members:
-    :show-inheritance:
