--- doc/source/data.rst
@@ -1,38 +1,0 @@
-data package
-============
-
-Subpackages
------------
-
-.. toctree::
-
-    data.fields
-    data.token_indexers
-
-Submodules
-----------
-
-data.instance module
---------------------
-
-.. automodule:: data.instance
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-data.vocabulary module
-----------------------
-
-.. automodule:: data.vocabulary
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: data
-    :members:
-    :undoc-members:
-    :show-inheritance:
