--- doc/source/data.token_indexers.rst
@@ -1,22 +1,0 @@
-data.token\_indexers package
-============================
-
-Submodules
-----------
-
-data.token\_indexers.token\_indexer module
-------------------------------------------
-
-.. automodule:: data.token_indexers.token_indexer
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: data.token_indexers
-    :members:
-    :undoc-members:
-    :show-inheritance:
