--- doc/source/index.rst
@@ -1,23 +1,0 @@
-.. pyAnt documentation master file, created by
-   sphinx-quickstart on Thu Nov 29 16:27:05 2018.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to pyAnt's documentation!
-=================================
-
-.. toctree::
-   :maxdepth: 3
-   :caption: Contents:
-
-API
-===
-.. automodule:: data
-   :members: instance
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
