--- doc/source/modules.rst
@@ -1,7 +1,0 @@
-pyAnt
-=====
-
-.. toctree::
-   :maxdepth: 4
-
-   data
