--- doc/source/pyAnt.data.dataset_readers.rst
@@ -1,22 +1,0 @@
-pyAnt.data.dataset\_readers package
-===================================
-
-Submodules
-----------
-
-pyAnt.data.dataset\_readers.dataset\_reader module
---------------------------------------------------
-
-.. automodule:: pyAnt.data.dataset_readers.dataset_reader
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: pyAnt.data.dataset_readers
-    :members:
-    :undoc-members:
-    :show-inheritance:
