--- doc/source/pyAnt.data.datasets.rst
@@ -1,22 +1,0 @@
-pyAnt.data.datasets package
-===========================
-
-Submodules
-----------
-
-pyAnt.data.datasets.dataset module
-----------------------------------
-
-.. automodule:: pyAnt.data.datasets.dataset
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: pyAnt.data.datasets
-    :members:
-    :undoc-members:
-    :show-inheritance:
