--- doc/source/pyAnt.data.fields.rst
@@ -1,38 +1,0 @@
-pyAnt.data.fields package
-=========================
-
-Submodules
-----------
-
-pyAnt.data.fields.field module
-------------------------------
-
-.. automodule:: pyAnt.data.fields.field
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-pyAnt.data.fields.index\_field module
--------------------------------------
-
-.. automodule:: pyAnt.data.fields.index_field
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-pyAnt.data.fields.text\_field module
-------------------------------------
-
-.. automodule:: pyAnt.data.fields.text_field
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: pyAnt.data.fields
-    :members:
-    :undoc-members:
-    :show-inheritance:
