--- doc/source/pyAnt.data.rst
@@ -1,40 +1,0 @@
-pyAnt.data package
-==================
-
-Subpackages
------------
-
-.. toctree::
-
-    pyAnt.data.dataset_readers
-    pyAnt.data.datasets
-    pyAnt.data.fields
-    pyAnt.data.token_indexers
-
-Submodules
-----------
-
-pyAnt.data.instance module
---------------------------
-
-.. automodule:: pyAnt.data.instance
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-pyAnt.data.vocabulary module
-----------------------------
-
-.. automodule:: pyAnt.data.vocabulary
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: pyAnt.data
-    :members:
-    :undoc-members:
-    :show-inheritance:
