--- doc/source/pyAnt.data.token_indexers.rst
@@ -1,38 +1,0 @@
-pyAnt.data.token\_indexers package
-==================================
-
-Submodules
-----------
-
-pyAnt.data.token\_indexers.char\_token\_indexer module
-------------------------------------------------------
-
-.. automodule:: pyAnt.data.token_indexers.char_token_indexer
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-pyAnt.data.token\_indexers.single\_id\_token\_indexer module
-------------------------------------------------------------
-
-.. automodule:: pyAnt.data.token_indexers.single_id_token_indexer
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-pyAnt.data.token\_indexers.token\_indexer module
-------------------------------------------------
-
-.. automodule:: pyAnt.data.token_indexers.token_indexer
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: pyAnt.data.token_indexers
-    :members:
-    :undoc-members:
-    :show-inheritance:
