--- doc/source/pyAnt.nn.dynet.rst
@@ -1,30 +1,0 @@
-pyAnt.nn.dynet package
-======================
-
-Submodules
-----------
-
-pyAnt.nn.dynet.multi\_layer\_perception module
-----------------------------------------------
-
-.. automodule:: pyAnt.nn.dynet.multi_layer_perception
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-pyAnt.nn.dynet.rnn\_builder module
-----------------------------------
-
-.. automodule:: pyAnt.nn.dynet.rnn_builder
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: pyAnt.nn.dynet
-    :members:
-    :undoc-members:
-    :show-inheritance:
