+++ setup.py
@@ -10,11 +10,11 @@ def read_file(fname):
     return open(os.path.join(os.path.dirname(__file__), fname)).read()
 
 setup(
-    name='pyAnt',
+    name='antu',
     version='0.0.1',
     author='AntNLP',
     author_email='taoji.cs@gmail.com',
-    description='Universal data IO module in NLP tasks (for AntNLP Group)',
+    description='Universal data IO and neural network modules in NLP tasks',
     long_description = read_file("README.txt"),
     license='Apache',
     packages=find_packages(),
@@ -25,7 +25,7 @@ setup(
         'Topic :: Documentation :: Sphinx',
         'Topic :: Scientific/Engineering :: Artificial Intelligence',
     ],
-    url ="https://github.com/AntNLP/pyAnt",
+    url ="https://github.com/AntNLP/antu",
     zip_safe=True,
     include_package_data=True,
     platforms='any',
